reimplement all routes with a trie
The prior hashmap is replaced with a trie which allows for greater
flexibility and lower time complexity. Code here isn't all that
clean, but functionality is roughly identical to before.

I say "roughly identical" because the API will now show all matches
for pronouns rather than picking the first at the whims of the HashMap.