implement wildcard matching
Routes like they/.../themself now work.

However, due to me being lazy, note that routes like
they/.../their/.../themself do not work as they make the whole
thing a lot harder. The algorithm for this isn't the prettiest
solution, but it's a step up from just matching the route, so I'm
okay with it.