trie: clean up guess_strings type signature
guess_strings has a kind of tricky return value because it must be
capable of building up PronounSet's incrementally. That is, it has
to store parts of PronounSets that are not themselves PronounSets.
The type signature did not make that clear. This commit extracts
out a newly documented type that hopefully makes the code a bit
clearer.

Also, now that the constraints on an IntPron are better expressed,
we can enforce them a bit better with an unwrap instead of an
unwrap_or.